Add architecture quality grading pass
