+++ Sources/QuillCodeApp/QuillCodeDesignSystem.swift
@@ -21,247 +21,6 @@ public enum QuillCodeMetrics {
     public static let pressScale: CGFloat = 0.96
 }
 
-struct QuillCodeHitTargetSpec {
-    enum Shape {
-        case rectangle
-        case rounded(CGFloat)
-        case capsule
-    }
-
-    var kind: QuillCodeNativeHitTargetKind
-    var action: String { kind.action.rawValue }
-    var allowsNestedInteractiveChildren: Bool { kind.allowsNestedInteractiveChildren }
-    var requiresUnblockedInterior: Bool { kind.requiresUnblockedInterior }
-    var requiresTactileFeedback: Bool { kind.requiresTactileFeedback }
-    var allowsTextSelection: Bool { kind.allowsTextSelection }
-    var minWidth: CGFloat?
-    var maxWidth: CGFloat?
-    var width: CGFloat?
-    var minHeight: CGFloat
-    var height: CGFloat?
-    var alignment: Alignment
-    var shape: Shape
-
-    private init(
-        kind: QuillCodeNativeHitTargetKind,
-        minWidth: CGFloat?,
-        maxWidth: CGFloat?,
-        width: CGFloat?,
-        minHeight: CGFloat,
-        height: CGFloat?,
-        alignment: Alignment,
-        shape: Shape
-    ) {
-        self.kind = kind
-        self.minWidth = minWidth.map(Self.clampedDimension) ?? QuillCodeMetrics.minimumHitTarget
-        self.maxWidth = maxWidth
-        self.width = width.map(Self.clampedDimension)
-        self.minHeight = Self.clampedDimension(minHeight)
-        self.height = height.map(Self.clampedDimension)
-        self.alignment = alignment
-        self.shape = shape
-    }
-
-    var requiredMinWidth: CGFloat {
-        width ?? minWidth ?? QuillCodeMetrics.minimumHitTarget
-    }
-
-    var requiredMinHeight: CGFloat {
-        height ?? minHeight
-    }
-
-    private static func clampedDimension(_ value: CGFloat) -> CGFloat {
-        max(value, QuillCodeMetrics.minimumHitTarget)
-    }
-
-    static func icon(
-        size: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        radius: CGFloat = QuillCodeMetrics.iconControlRadius
-    ) -> Self {
-        Self(
-            kind: .icon,
-            minWidth: size,
-            maxWidth: nil,
-            width: size,
-            minHeight: QuillCodeMetrics.minimumHitTarget,
-            height: size,
-            alignment: .center,
-            shape: .rounded(radius)
-        )
-    }
-
-    static func textButton(
-        minWidth: CGFloat = QuillCodeMetrics.compactTextButtonMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> Self {
-        Self(
-            kind: .textButton,
-            minWidth: minWidth,
-            maxWidth: nil,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rounded(radius)
-        )
-    }
-
-    static func formAction(
-        minWidth: CGFloat = QuillCodeMetrics.compactFormActionMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> Self {
-        Self(
-            kind: .formAction,
-            minWidth: minWidth,
-            maxWidth: nil,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rounded(QuillCodeMetrics.minimumHitTarget / 2)
-        )
-    }
-
-    static func textEntry(
-        minWidth: CGFloat? = nil,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> Self {
-        Self(
-            kind: .textEntry,
-            minWidth: minWidth,
-            maxWidth: nil,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rounded(radius)
-        )
-    }
-
-    static func segmentedControl(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> Self {
-        Self(
-            kind: .segmentedControl,
-            minWidth: nil,
-            maxWidth: .infinity,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rounded(QuillCodeMetrics.compactControlRadius)
-        )
-    }
-
-    static func adjustableControl(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> Self {
-        Self(
-            kind: .adjustableControl,
-            minWidth: nil,
-            maxWidth: .infinity,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rounded(QuillCodeMetrics.compactControlRadius)
-        )
-    }
-
-    static func link(
-        minWidth: CGFloat? = QuillCodeMetrics.compactTextButtonMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> Self {
-        Self(
-            kind: .link,
-            minWidth: minWidth,
-            maxWidth: nil,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rounded(radius)
-        )
-    }
-
-    static func switchRow(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading
-    ) -> Self {
-        Self(
-            kind: .switchRow,
-            minWidth: nil,
-            maxWidth: .infinity,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rectangle
-        )
-    }
-
-    static func ownedGesture(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> Self {
-        Self(
-            kind: .ownedGesture,
-            minWidth: nil,
-            maxWidth: .infinity,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rounded(radius)
-        )
-    }
-
-    static func fullRow(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> Self {
-        Self(
-            kind: .fullRow,
-            minWidth: nil,
-            maxWidth: .infinity,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .rounded(radius)
-        )
-    }
-
-    static func capsule(
-        minWidth: CGFloat? = nil,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> Self {
-        Self(
-            kind: .capsule,
-            minWidth: minWidth,
-            maxWidth: nil,
-            width: nil,
-            minHeight: minHeight,
-            height: nil,
-            alignment: alignment,
-            shape: .capsule
-        )
-    }
-
-}
-
 enum QuillCodePalette {
     static let background = Color(red: 0.03, green: 0.06, blue: 0.08)
     static let sidebar = Color(red: 0.07, green: 0.10, blue: 0.12)
@@ -387,149 +146,6 @@ public struct QuillCodeActionButtonStyle: ButtonStyle {
 }
 
 extension View {
-    fileprivate func quillCodeInteractiveTarget(_ spec: QuillCodeHitTargetSpec) -> some View {
-        modifier(QuillCodeHitTargetModifier(spec: spec))
-    }
-
-    public func quillCodeTextButtonTarget(
-        minWidth: CGFloat = QuillCodeMetrics.compactTextButtonMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.textButton(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-    }
-
-    public func quillCodeFormActionTarget(
-        minWidth: CGFloat = QuillCodeMetrics.compactFormActionMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> some View {
-        quillCodeInteractiveTarget(.formAction(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeTextEntryTarget(
-        minWidth: CGFloat? = nil,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.textEntry(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-    }
-
-    public func quillCodeSegmentedControlTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> some View {
-        quillCodeInteractiveTarget(.segmentedControl(
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeAdjustableControlTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> some View {
-        quillCodeInteractiveTarget(.adjustableControl(
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeLinkTarget(
-        minWidth: CGFloat? = QuillCodeMetrics.compactTextButtonMinWidth,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.link(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-    }
-
-    public func quillCodeSwitchRowTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading
-    ) -> some View {
-        quillCodeInteractiveTarget(.switchRow(
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeOwnedGestureTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.ownedGesture(
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-        .accessibilityAddTraits(.isButton)
-    }
-
-    public func quillCodeIconButtonTarget(
-        size: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        radius: CGFloat = QuillCodeMetrics.iconControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.icon(size: size, radius: radius))
-    }
-
-    public func quillCodeFullRowButtonTarget(
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .leading,
-        radius: CGFloat = QuillCodeMetrics.compactControlRadius
-    ) -> some View {
-        quillCodeInteractiveTarget(.fullRow(
-            minHeight: minHeight,
-            alignment: alignment,
-            radius: radius
-        ))
-    }
-
-    public func quillCodeCapsuleButtonTarget(
-        minWidth: CGFloat? = nil,
-        minHeight: CGFloat = QuillCodeMetrics.minimumHitTarget,
-        alignment: Alignment = .center
-    ) -> some View {
-        quillCodeInteractiveTarget(.capsule(
-            minWidth: minWidth,
-            minHeight: minHeight,
-            alignment: alignment
-        ))
-    }
-
-    public func quillCodeDecorativeIconFrame(
-        size: CGFloat = QuillCodeMetrics.minimumHitTarget
-    ) -> some View {
-        frame(width: size, height: size)
-            .accessibilityHidden(true)
-    }
-
-    public func quillCodePlatformMenuItemTarget(reason: String) -> some View {
-        modifier(QuillCodePlatformMenuItemTargetModifier(reason: reason))
-    }
-
     func quillCodeSurface(
         fill: Color,
         radius: CGFloat,
@@ -546,55 +162,6 @@ extension View {
 
     func quillCodeImageOutline(radius: CGFloat) -> some View {
         modifier(QuillCodeImageOutlineModifier(radius: radius))
-    }
-}
-
-private struct QuillCodeHitTargetModifier: ViewModifier {
-    var spec: QuillCodeHitTargetSpec
-
-    func body(content: Content) -> some View {
-        shaped(content)
-    }
-
-    @ViewBuilder
-    private func shaped(_ content: Content) -> some View {
-        switch spec.shape {
-        case .rectangle:
-            framed(content)
-                .contentShape(Rectangle())
-        case .rounded(let radius):
-            framed(content)
-                .contentShape(RoundedRectangle(cornerRadius: radius, style: .continuous))
-        case .capsule:
-            framed(content)
-                .contentShape(Capsule())
-        }
-    }
-
-    private func framed(_ content: Content) -> some View {
-        content.frame(
-            minWidth: requiredMinWidth,
-            maxWidth: spec.maxWidth,
-            minHeight: requiredMinHeight,
-            alignment: spec.alignment
-        )
-        .frame(width: width, height: height, alignment: spec.alignment)
-    }
-
-    private var requiredMinWidth: CGFloat {
-        max(spec.requiredMinWidth, QuillCodeMetrics.minimumHitTarget)
-    }
-
-    private var requiredMinHeight: CGFloat {
-        max(spec.requiredMinHeight, QuillCodeMetrics.minimumHitTarget)
-    }
-
-    private var width: CGFloat? {
-        spec.width.map { max($0, QuillCodeMetrics.minimumHitTarget) }
-    }
-
-    private var height: CGFloat? {
-        spec.height.map { max($0, QuillCodeMetrics.minimumHitTarget) }
     }
 }
 
@@ -631,14 +198,5 @@ private struct QuillCodeImageOutlineModifier: ViewModifier {
                 RoundedRectangle(cornerRadius: radius, style: .continuous)
                     .stroke(Color.white.opacity(0.10), lineWidth: 1)
             )
-    }
-}
-
-private struct QuillCodePlatformMenuItemTargetModifier: ViewModifier {
-    var reason: String
-
-    func body(content: Content) -> some View {
-        _ = reason
-        return content
     }
 }
