+++ Tests/QuillCodeParityTests/ParityAutomationGateTests.swift
@@ -23,9 +23,11 @@ final class ParityAutomationGateTests: QuillCodeParityTestCase {
     func testWorkspaceModelDelegatesAutomationStateMutations() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
         let modelAutomationText = try Self.appSourceText(named: "WorkspaceModelAutomations.swift")
+        let modelAutomationRunText = try Self.appSourceText(named: "WorkspaceModelAutomationRuns.swift")
         let automationText = try Self.appSourceText(named: "WorkspaceAutomationEngine.swift")
 
         XCTAssertTrue(modelAutomationText.contains("extension QuillCodeWorkspaceModel"), "Automation model API should live in a focused workspace model extension.")
+        XCTAssertTrue(modelAutomationRunText.contains("extension QuillCodeWorkspaceModel"), "Automation run API should live in a focused workspace model extension.")
         XCTAssertTrue(automationText.contains("enum WorkspaceAutomationStateReducer"), "Automation state mutation should live in a focused reducer.")
         XCTAssertTrue(automationText.contains("struct WorkspaceAutomationStateMutation"), "Automation state mutations should return typed mutation results.")
         XCTAssertTrue(automationText.contains("static func setItems"), "Automation sorting and visibility preservation should be reducer-owned.")
@@ -37,13 +39,16 @@ final class ParityAutomationGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.setItems"), "WorkspaceModel automation extension should delegate automation item setting.")
         XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.createThreadFollowUp"), "WorkspaceModel automation extension should delegate thread follow-up creation.")
         XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.createWorkspaceSchedule"), "WorkspaceModel automation extension should delegate workspace schedule creation.")
-        XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.updateStatus"), "WorkspaceModel automation extension should delegate status changes.")
-        XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.delete"), "WorkspaceModel automation extension should delegate deletion.")
-        XCTAssertTrue(modelAutomationText.contains("WorkspaceAutomationStateReducer.replace"), "WorkspaceModel automation extension should delegate replacement.")
+        XCTAssertTrue(modelAutomationRunText.contains("WorkspaceAutomationStateReducer.updateStatus"), "WorkspaceModel automation run extension should delegate status changes.")
+        XCTAssertTrue(modelAutomationRunText.contains("WorkspaceAutomationStateReducer.delete"), "WorkspaceModel automation run extension should delegate deletion.")
+        XCTAssertTrue(modelAutomationRunText.contains("WorkspaceAutomationStateReducer.replace"), "WorkspaceModel automation run extension should delegate replacement.")
         XCTAssertTrue(automationText.contains("struct WorkspaceAutomationTrigger"), "Due automation execution should preserve trigger metadata.")
         XCTAssertTrue(automationText.contains("static func dueAutomationTriggers"), "Trigger-aware due selection should stay in the automation runner.")
-        XCTAssertTrue(modelAutomationText.contains("automationEventSources()"), "WorkspaceModel automation extension should resolve monitor event sources near run orchestration.")
-        XCTAssertTrue(modelAutomationText.contains("eventDescription:"), "WorkspaceModel automation runs should pass event-source trigger descriptions into monitor drafts.")
+        XCTAssertTrue(modelAutomationRunText.contains("automationEventSources()"), "WorkspaceModel automation run extension should resolve monitor event sources near run orchestration.")
+        XCTAssertTrue(modelAutomationRunText.contains("eventDescription:"), "WorkspaceModel automation runs should pass event-source trigger descriptions into monitor drafts.")
+        XCTAssertFalse(modelAutomationText.contains("WorkspaceAutomationStateReducer.updateStatus"), "WorkspaceModel automation creation extension should not own status changes.")
+        XCTAssertFalse(modelAutomationText.contains("WorkspaceAutomationStateReducer.delete"), "WorkspaceModel automation creation extension should not own deletion.")
+        XCTAssertFalse(modelAutomationText.contains("WorkspaceAutomationStateReducer.replace"), "WorkspaceModel automation creation extension should not own replacement.")
         XCTAssertFalse(modelText.contains("public func createThreadFollowUpAutomation"), "WorkspaceModel.swift should not own automation scheduling APIs.")
         XCTAssertFalse(modelText.contains("public func createWorkspaceScheduleAutomation"), "WorkspaceModel.swift should not own workspace-check scheduling APIs.")
         XCTAssertFalse(modelText.contains("public func runDueAutomations"), "WorkspaceModel.swift should not own automation-run orchestration.")
