+++ Tests/QuillCodeParityTests/ParityWorkspaceSidebarGateTests.swift
@@ -145,6 +145,8 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
         let surfaceText = try Self.appSourceText(named: "WorkspaceSurface.swift")
         let projectListText = try Self.appSourceText(named: "QuillCodeProjectListSurface.swift")
         let sidebarText = try Self.appSourceText(named: "QuillCodeThreadSidebarSurface.swift")
+        let sidebarItemText = try Self.appSourceText(named: "QuillCodeSidebarItemSurface.swift")
+        let sidebarFilterText = try Self.appSourceText(named: "QuillCodeSidebarFilterSurface.swift")
         let threadListBuilderText = try Self.appSourceText(named: "QuillCodeSidebarThreadListBuilder.swift")
         let broadSurfaceTests = try Self.appTestSourceText(named: "WorkspaceSurfaceTests.swift")
         let sidebarSurfaceTests = try Self.appTestSourceText(named: "QuillCodeThreadSidebarSurfaceTests.swift")
@@ -155,11 +157,13 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(projectListText.contains("public enum ProjectItemActionKind"), "Project action labels should live in project-list contracts.")
         XCTAssertTrue(projectListText.contains("public struct ProjectItemActionSurface"), "Project action records should live in project-list contracts.")
         XCTAssertTrue(sidebarText.contains("public struct SidebarSurface"), "Thread sidebar aggregate records should live in thread-sidebar contracts.")
-        XCTAssertTrue(sidebarText.contains("public struct SidebarItemSurface"), "Thread sidebar item rows should live in thread-sidebar contracts.")
-        XCTAssertTrue(sidebarText.contains("public enum SidebarBulkActionKind"), "Thread bulk action labels should live in thread-sidebar contracts.")
-        XCTAssertTrue(sidebarText.contains("public struct SidebarBulkActionSurface"), "Thread bulk action command IDs should live in thread-sidebar contracts.")
-        XCTAssertTrue(sidebarText.contains("public enum SidebarItemActionKind"), "Thread action labels should live in thread-sidebar contracts.")
-        XCTAssertTrue(sidebarText.contains("public struct SidebarItemActionSurface"), "Thread action records should live in thread-sidebar contracts.")
+        XCTAssertTrue(sidebarItemText.contains("public struct SidebarItemSurface"), "Thread sidebar item rows should live in focused item contracts.")
+        XCTAssertTrue(sidebarItemText.contains("public enum SidebarBulkActionKind"), "Thread bulk action labels should live in focused item contracts.")
+        XCTAssertTrue(sidebarItemText.contains("public struct SidebarBulkActionSurface"), "Thread bulk action command IDs should live in focused item contracts.")
+        XCTAssertTrue(sidebarItemText.contains("public enum SidebarItemActionKind"), "Thread action labels should live in focused item contracts.")
+        XCTAssertTrue(sidebarItemText.contains("public struct SidebarItemActionSurface"), "Thread action records should live in focused item contracts.")
+        XCTAssertTrue(sidebarFilterText.contains("public enum SidebarSavedFilterKind"), "Thread saved filters should live in focused filter contracts.")
+        XCTAssertTrue(sidebarFilterText.contains("public struct SidebarSavedSearchSurface"), "Thread saved searches should live in focused filter contracts.")
         XCTAssertTrue(sidebarText.contains("filteredItems"), "Sidebar search filtering should be directly testable outside the aggregate workspace surface.")
         XCTAssertTrue(sidebarText.contains("selectionLabel"), "Sidebar selection copy should be directly testable outside the aggregate workspace surface.")
         XCTAssertTrue(sidebarText.contains("SidebarThreadListBuilder(items: items)"), "Sidebar aggregate should delegate thread list derivation.")
@@ -171,6 +175,8 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
         XCTAssertFalse(sidebarText.contains("public struct ProjectListSurface"), "Thread-sidebar contracts should not own project list records.")
         XCTAssertFalse(sidebarText.contains("public struct ProjectItemSurface"), "Thread-sidebar contracts should not own project rows.")
         XCTAssertFalse(sidebarText.contains("ProjectItemActionSurface"), "Thread-sidebar contracts should not own project actions.")
+        XCTAssertFalse(sidebarItemText.contains("public struct SidebarSurface"), "Thread item contracts should not own aggregate sidebar records.")
+        XCTAssertFalse(sidebarFilterText.contains("public struct SidebarSurface"), "Thread filter contracts should not own aggregate sidebar records.")
         XCTAssertFalse(surfaceText.contains("public struct ProjectListSurface"), "WorkspaceSurface should not own project list surface records.")
         XCTAssertFalse(surfaceText.contains("public struct ProjectItemSurface"), "WorkspaceSurface should not own project row records.")
         XCTAssertFalse(surfaceText.contains("public enum ProjectItemActionKind"), "WorkspaceSurface should not own project action labels.")
